edge_device/requirements.txt
# RDK X5 Dependencies
cv2-python==4.8.1.78
numpy==1.24.3
requests==2.31.0

# Note: hobot_dnn and libsrcampy are pre-installed on RDK X5 system
# These are proprietary Horizon Robotics libraries
